MDEV-15373 post-push: fixing type mismatch alerted by windows build.

+++ sql/rpl_rli.cc
@@ -1780,7 +1780,8 @@ gtid_pos_auto_create_tables(rpl_slave_state::gtid_pos_table **list_ptr)
     p= strmake(p, plugin_name(*auto_engines)->str, FN_REFLEN - (p - buf));
     table_name.str= buf;
     table_name.length= p - buf;
-    table_case_convert(const_cast<char*>(table_name.str), table_name.length);
+    table_case_convert(const_cast<char*>(table_name.str),
+                       static_cast<uint>(table_name.length));
     entry= rpl_global_gtid_slave_state->alloc_gtid_pos_table
       (&table_name, hton, rpl_slave_state::GTID_POS_AUTO_CREATE);
     if (!entry)
